-changed folders\files organizations. Added Common folder to contain files used by both projects
-many minor changed due to update of incude file paths

- changed representation of brightness level value in csv files
- fixed bug in apply method of filter

-other minor changes

+++ src/fgt/edit_quality/common/const_types.h
@@ -36,7 +36,7 @@ struct EQUALIZER_INFO
 	float minQualityVal;
 	float midQualityPercentage;
 	float maxQualityVal;
-	int brightness;
+	float brightness;
 };
 
 #endif
